refactor(activerecord): converge JoinDependency constructor onto Rails' arg order
JoinDependency's constructor now takes `(model, table, associations,
joinType)`, matching `JoinDependency#initialize` (join_dependency.rb:71),
and builds the association tree in the constructor like Rails does.

The duplicate `walkAssociationTree` / `addTreeToJoinDependency` pair in
query-methods.ts is folded into the class: tree normalization goes through
the existing `JoinDependency.makeTree` / `walkTree` statics (which now
raise ConfigurationError on an invalid spec, mirroring walk_tree's else
branch) and the node build into a private `_buildTree` mirroring
`JoinDependency#build`. `constructJoinDependency` is now the one-liner
Rails has.
